rfr: refactor
