+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
         android:theme="@style/Theme.AutoGestion"
         tools:targetApi="31">
         <activity
-            android:name=".Home"
+            android:name=".ui.Home"
             android:exported="true"
             android:theme="@style/Theme.AutoGestion">
             <intent-filter>
@@ -37,17 +37,17 @@
         </activity>
 
         <!-- <activity android:name=".Home"/> -->
-        <activity android:name=".Camera" />
-        <activity android:name=".ClientProfile" />
-        <activity android:name=".VehicleProfile" />
-        <activity android:name=".form.ClientForm" />
-        <activity android:name=".form.VehicleForm" />
-        <activity android:name=".form.RepairForm" />
-        <activity android:name=".form.ClientFormUpdate" />
-        <activity android:name=".form.VehicleFormUpdate" />
-        <activity android:name=".form.RepairFormUpdate" />
-        <activity android:name=".form.VehicleFormAdd" />
-        <activity android:name=".form.RepairFormAdd" />
+        <activity android:name=".ui.components.Camera" />
+        <activity android:name=".ui.profiles.ClientProfile" />
+        <activity android:name=".ui.profiles.VehicleProfile" />
+        <activity android:name=".ui.form.ClientForm" />
+        <activity android:name=".ui.form.VehicleForm" />
+        <activity android:name=".ui.form.RepairForm" />
+        <activity android:name=".ui.form.ClientFormUpdate" />
+        <activity android:name=".ui.form.VehicleFormUpdate" />
+        <activity android:name=".ui.form.RepairFormUpdate" />
+        <activity android:name=".ui.form.VehicleFormAdd" />
+        <activity android:name=".ui.form.RepairFormAdd" />
 
         <provider
             android:name="androidx.core.content.FileProvider"
