+++ app/src/main/java/com/example/autogestion/ui/form/ClientForm.kt
@@ -1,4 +1,4 @@
-package com.example.autogestion.form
+package com.example.autogestion.ui.form
 
 import android.content.Intent
 import android.os.Bundle
@@ -16,8 +16,8 @@ import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.autogestion.Home
-import com.example.autogestion.NavBar
+import com.example.autogestion.ui.Home
+import com.example.autogestion.ui.components.NavBar
 import com.example.autogestion.data.Client
 import com.example.autogestion.data.viewModels.ClientViewModel
 import kotlinx.coroutines.launch
