+++ app/src/main/java/com/example/autogestion/ui/form/ClientFormUpdate.kt
@@ -1,4 +1,4 @@
-package com.example.autogestion.form
+package com.example.autogestion.ui.form
 
 import android.app.DatePickerDialog
 import android.content.Intent
@@ -17,13 +17,11 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.autogestion.ClientProfile
-import com.example.autogestion.Home
-import com.example.autogestion.NavBar
-import com.example.autogestion.VehicleProfile
+import com.example.autogestion.ui.profiles.ClientProfile
+import com.example.autogestion.ui.Home
+import com.example.autogestion.ui.components.NavBar
 import com.example.autogestion.data.Client
 import com.example.autogestion.data.viewModels.ClientViewModel
 import kotlinx.coroutines.launch
