+++ app/src/main/java/com/example/autogestion/ui/form/RepairFormUpdate.kt
@@ -1,4 +1,4 @@
-package com.example.autogestion.form
+package com.example.autogestion.ui.form
 
 import android.content.Intent
 import android.net.Uri
@@ -21,8 +21,8 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.autogestion.NavBar
-import com.example.autogestion.VehicleProfile
+import com.example.autogestion.ui.components.NavBar
+import com.example.autogestion.ui.profiles.VehicleProfile
 import com.example.autogestion.data.Repair
 import com.example.autogestion.data.viewModels.RepairViewModel
 import kotlinx.coroutines.launch
@@ -30,7 +30,7 @@ import java.text.ParseException
 import java.text.SimpleDateFormat
 import java.util.Calendar
 import java.util.Locale
-import com.example.autogestion.Utils.getFilePathFromUri
+import com.example.autogestion.ui.utils.getFilePathFromUri
 
 class RepairFormUpdate : ComponentActivity() {
