+++ app/src/main/java/com/example/autogestion/ui/utils/FileUtils.kt
@@ -1,4 +1,4 @@
-package com.example.autogestion.Utils
+package com.example.autogestion.ui.utils
 
 import android.content.Context
 import android.net.Uri
